Setting up production level project structure; inorder to make code modular
