+++ package.json
@@ -12,6 +12,8 @@
   "license": "ISC",
   "dependencies": {
     "express": "^4.17.1",
-    "nodemon": "^2.0.15"
+    "jest": "^27.4.3",
+    "nodemon": "^2.0.15",
+    "winston": "^3.3.3"
   }
 }
